+++ example/network/sockets/udp_multicast/sdkconfig.h
@@ -47,7 +47,6 @@
 /* CONFIG_GCC_CODE_MODEL_LARGE is not set */
 /* end of Compiler configuration */
 #define CONFIG_USE_CACHE
-/* CONFIG_USE_L3CACHE is not set */
 #define CONFIG_USE_MMU
 /* CONFIG_BOOT_WITH_FLUSH_CACHE is not set */
 /* CONFIG_MMU_DEBUG_PRINTS is not set */
@@ -56,18 +55,19 @@
 
 /* Soc configuration */
 
-/* CONFIG_TARGET_PHYTIUMPI is not set */
+#define CONFIG_TARGET_PHYTIUMPI
 /* CONFIG_TARGET_E2000Q is not set */
 /* CONFIG_TARGET_E2000D is not set */
 /* CONFIG_TARGET_E2000S is not set */
 /* CONFIG_TARGET_FT2004 is not set */
-#define CONFIG_TARGET_D2000
-#define CONFIG_SOC_NAME "d2000"
-#define CONFIG_SOC_CORE_NUM 8
+/* CONFIG_TARGET_D2000 is not set */
+#define CONFIG_SOC_NAME "phytiumpi"
+#define CONFIG_SOC_CORE_NUM 4
 #define CONFIG_F32BIT_MEMORY_ADDRESS 0x80000000
 #define CONFIG_F32BIT_MEMORY_LENGTH 0x80000000
 #define CONFIG_F64BIT_MEMORY_ADDRESS 0x2000000000
 #define CONFIG_F64BIT_MEMORY_LENGTH 0x800000000
+#define CONFIG_TARGET_E2000
 /* CONFIG_USE_SPINLOCK is not set */
 #define CONFIG_DEFAULT_DEBUG_PRINT_UART1
 /* CONFIG_DEFAULT_DEBUG_PRINT_UART0 is not set */
@@ -76,11 +76,21 @@
 
 /* Board Configuration */
 
-#define CONFIG_BOARD_NAME "test"
-#define CONFIG_D2000_TEST_BOARD
+#define CONFIG_BOARD_NAME "firefly"
+/* CONFIG_USE_SPI_IOPAD is not set */
+/* CONFIG_USE_GPIO_IOPAD is not set */
+/* CONFIG_USE_CAN_IOPAD is not set */
+/* CONFIG_USE_QSPI_IOPAD is not set */
+/* CONFIG_USE_PWM_IOPAD is not set */
+/* CONFIG_USE_MIO_IOPAD is not set */
+/* CONFIG_USE_TACHO_IOPAD is not set */
+/* CONFIG_USE_UART_IOPAD is not set */
+/* CONFIG_USE_THIRD_PARTY_IOPAD is not set */
+#define CONFIG_FIREFLY_DEMO_BOARD
 
 /* IO mux configuration when board start up */
 
+/* end of IO mux configuration when board start up */
 /* CONFIG_CUS_DEMO_BOARD is not set */
 
 /* Build project name */
@@ -113,8 +123,8 @@
 /* Drivers configuration */
 
 #define CONFIG_USE_IOMUX
-#define CONFIG_ENABLE_IOCTRL
-/* CONFIG_ENABLE_IOPAD is not set */
+/* CONFIG_ENABLE_IOCTRL is not set */
+#define CONFIG_ENABLE_IOPAD
 /* CONFIG_USE_SPI is not set */
 /* CONFIG_USE_QSPI is not set */
 #define CONFIG_USE_GIC
@@ -130,10 +140,10 @@
 
 /* Eth Configuration */
 
-/* CONFIG_ENABLE_FXMAC is not set */
-#define CONFIG_ENABLE_FGMAC
-#define CONFIG_FGMAC_PHY_COMMON
-/* CONFIG_FGMAC_PHY_AR803X is not set */
+#define CONFIG_ENABLE_FXMAC
+/* CONFIG_ENABLE_FGMAC is not set */
+#define CONFIG_FXMAC_PHY_COMMON
+/* CONFIG_FXMAC_PHY_YT is not set */
 /* end of Eth Configuration */
 /* CONFIG_USE_CAN is not set */
 /* CONFIG_USE_I2C is not set */
@@ -227,8 +237,8 @@
 
 /* Freertos Eth Drivers */
 
-/* CONFIG_FREERTOS_USE_XMAC is not set */
-#define CONFIG_FREERTOS_USE_GMAC
+#define CONFIG_FREERTOS_USE_XMAC
+/* CONFIG_FREERTOS_USE_GMAC is not set */
 /* end of Freertos Eth Drivers */
 
 /* Freertos Gpio Drivers */
@@ -288,8 +298,8 @@
 
 /* LWIP Port Configuration */
 
-/* CONFIG_LWIP_FXMAC is not set */
-#define CONFIG_LWIP_FGMAC
+#define CONFIG_LWIP_FXMAC
+/* CONFIG_LWIP_FGMAC is not set */
 /* CONFIG_LWIP_RX_POLL is not set */
 /* end of LWIP Port Configuration */
 /* CONFIG_LWIP_NO_SYS is not set */