+++ example/peripheral/spi/sdkconfig.h
@@ -10,31 +10,32 @@
 
 /* Arm architecture configuration */
 
-/* CONFIG_ARCH_ARMV8_AARCH64 is not set */
-#define CONFIG_ARCH_ARMV8_AARCH32
+#define CONFIG_ARCH_ARMV8_AARCH64
+/* CONFIG_ARCH_ARMV8_AARCH32 is not set */
 
 /* Compiler configuration */
 
 #define CONFIG_ARM_GCC_SELECT
 /* CONFIG_ARM_CLANG_SELECT is not set */
 #define CONFIG_TOOLCHAIN_NAME "gcc"
-#define CONFIG_TARGET_ARMV8_AARCH32
-#define CONFIG_ARCH_EXECUTION_STATE "aarch32"
+#define CONFIG_TARGET_ARMV8_AARCH64
+#define CONFIG_ARCH_EXECUTION_STATE "aarch64"
 
 /* Fpu configuration */
 
-/* CONFIG_ARCH_FPU_VFP_V3 is not set */
-#define CONFIG_ARCH_FPU_VFP_V4
-#define CONFIG_ARM_DPFPU32
-/* CONFIG_ARM_FPU_ABI_SOFT is not set */
-#define CONFIG_ARM_FPU_SYMBOL "crypto-neon-fp-armv8"
 #define CONFIG_ARM_NEON
-#define CONFIG_ARCH_FPU
 /* end of Fpu configuration */
+#define CONFIG_ARM_CRC
+#define CONFIG_ARM_CRYPTO
+#define CONFIG_ARM_FLOAT_POINT
+/* CONFIG_GCC_CODE_MODEL_TINY is not set */
+#define CONFIG_GCC_CODE_MODEL_SMALL
+/* CONFIG_GCC_CODE_MODEL_LARGE is not set */
 /* end of Compiler configuration */
 #define CONFIG_USE_CACHE
 #define CONFIG_USE_MMU
-#define CONFIG_USE_AARCH64_L1_TO_AARCH32
+/* CONFIG_BOOT_WITH_FLUSH_CACHE is not set */
+/* CONFIG_MMU_DEBUG_PRINTS is not set */
 /* end of Arm architecture configuration */
 /* end of Arch configuration */
 
@@ -186,12 +187,8 @@
 #define CONFIG_IMAGE_LOAD_ADDRESS 0x80100000
 #define CONFIG_IMAGE_MAX_LENGTH 0x1000000
 #define CONFIG_HEAP_SIZE 1
-#define CONFIG_SVC_STACK_SIZE 0x1000
-#define CONFIG_SYS_STACK_SIZE 0x1000
-#define CONFIG_IRQ_STACK_SIZE 0x1000
-#define CONFIG_ABORT_STACK_SIZE 0x1000
-#define CONFIG_FIQ_STACK_SIZE 0x1000
-#define CONFIG_UNDEF_STACK_SIZE 0x1000
+#define CONFIG_STACK_SIZE 0x400
+#define CONFIG_FPU_STACK_SIZE 0x1000
 /* end of Linker Options */
 /* end of Build setup */